[BUGFIX] Update scope of norm with recent dates
The update in the REST API for LocalLaw provides the ability to replace older norms with newer ones in the normScope. In the function, a check is performed to compare the date of previously stored norms with the new ones. If the new date is more recent, the old norm is replaced, else the new norm is discarded.